+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [unreleased]
 
+### FEATURES
+
+- feat: add ccv-provider module in v8 upgrade ([\#558](https://github.com/MANTRA-Chain/mantrachain/issues/558))
+
 ## v7.0.0
 
 *December 29, 2025*